Buzzbrief/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
python-dotenv==1.0.1
requests==2.31.0
azure-cognitiveservices-speech==1.34.1  # Only include if you're using Azure TTS